fix: マイグレーションファイルの修正

+++ laravel-project/database/migrations/2024_07_12_055050_create_feedback_table.php
@@ -10,7 +10,7 @@ class CreateFeedbackTable extends Migration
     {
         Schema::create('feedback', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('idea_id');
+            $table->unsignedInteger('idea_id');
             $table->integer('fb_chart1');
             $table->integer('fb_chart2');
             $table->integer('fb_chart3');
